Add more gcc flags and fix warnings compilation (part 1)

+++ src/common/containers/Collection.ipp
@@ -22,18 +22,18 @@
 // **************************************************************
 //! \param elt is the element of type T to insert.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-void Collection<T,N,Block>::insert(const uint32_t nth, T const& elt)
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+void Collection<T,N,Block>::insert(const size_t nth, T const& elt)
 {
-  const uint32_t id = nth >> N;
-  const uint32_t sid = MODULO(nth, M);
+  const size_t id = nth >> N;
+  const size_t sid = MODULO(nth, M);
 
   // Reserve a new block of elements if all current blocks are
   // occupied.
   if (id >= IContainer<T,N,Block>::m_allocated_blocks)
     {
-      IContainer<T,N,Block>::reserveBlocks(1U + id - IContainer<T,N,Block>::m_allocated_blocks);
+      IContainer<T,N,Block>::reserveBlocks(1ul + id - IContainer<T,N,Block>::m_allocated_blocks);
     }
 
   // Insert element and add the 'Occupied' flag
@@ -54,8 +54,8 @@ void Collection<T,N,Block>::insert(const uint32_t nth, T const& elt)
 // **************************************************************
 //! \param elt is the element of type T to insert.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
 void Collection<T,N,Block>::insert(T const& elt)
 {
   insert(m_end, elt);
@@ -65,17 +65,17 @@ void Collection<T,N,Block>::insert(T const& elt)
 //! \param nth the n'th element (index) of the container we want
 //! to be not empty.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-void Collection<T,N,Block>::occupy(const uint32_t nth)
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+void Collection<T,N,Block>::occupy(const size_t nth)
 {
   if (outofbound(nth))
     {
       throw std::out_of_range("Out of range index " + std::to_string(nth));
     }
 
-  const uint32_t id = nth >> N;
-  const uint32_t sid = MODULO(nth, M);
+  const size_t id = nth >> N;
+  const size_t sid = MODULO(nth, M);
 
   if (!IS_OCCUPIED(id, sid))
     {
@@ -84,7 +84,7 @@ void Collection<T,N,Block>::occupy(const uint32_t nth)
 
       // Update min and max bounding box used for iterators
       if (nth >= m_end)
-        m_end = nth + 1;
+        m_end = nth + 1ul;
       if (nth < m_begin)
         m_begin = nth;
     }
@@ -94,15 +94,15 @@ void Collection<T,N,Block>::occupy(const uint32_t nth)
 //! \param nth the n'th element (index) of the container we want
 //! to remove.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-void Collection<T,N,Block>::remove(const uint32_t nth)
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+void Collection<T,N,Block>::remove(const size_t nth)
 {
   if (outofbound(nth))
     return ;
 
-  const uint32_t id = nth >> N;
-  const uint32_t sid = MODULO(nth, M);
+  const size_t id = nth >> N;
+  const size_t sid = MODULO(nth, M);
 
   // Remove the 'Occupied' flag
   if (IS_OCCUPIED(id, sid))
@@ -116,16 +116,16 @@ void Collection<T,N,Block>::remove(const uint32_t nth)
             ++m_begin;
           } while ((m_begin < m_end) && (!IContainer<T,N,Block>::occupied(m_begin)));
         }
-      else if (nth + 1U == m_end)
+      else if (nth + 1ul == m_end)
         {
-          while (m_end && (!IContainer<T,N,Block>::occupied(m_end - 1U)))
+          while (m_end && (!IContainer<T,N,Block>::occupied(m_end - 1)))
             {
               --m_end;
             }
         }
       if (m_end == m_begin)
         {
-          m_begin = (uint32_t) - 1;
+          m_begin = (size_t) -1;
           m_end = 0;
         }
       std::cout << "after removing ==> Min: " << m_begin << "  Max: " << m_end << "\n";
