+++ src/common/containers/IContainer.ipp
@@ -20,11 +20,11 @@
 //=====================================================================
 
 // Perform a % b but works if and only if M is a power of two number.
-#define MODULO(a, b)             ((a) & ((b) - 1U))
+#define MODULO(a, b)             ((a) & ((b) - 1ul))
 // Accessor to the bitfield indicationg which elements are stored.
 #define OCCUPIED(id, sid)        (IContainer<T,N,Block>::m_blocks[id]->m_occupied[sid / S])
 // Perform a generic operation on the bitfield.
-#define OP_OCCUPIED(op, id, sid) (OCCUPIED(id, sid) op (1 << (MODULO(sid, S))))
+#define OP_OCCUPIED(op, id, sid) (OCCUPIED(id, sid) op (1ul << (MODULO(sid, S))))
 // Set the 'occupied' flag in the bitfield.
 #define SET_OCCUPIED(id, sid)    (OP_OCCUPIED(|=, id, sid))
 // Set the 'empty' flag in the bitfield.
@@ -38,9 +38,9 @@
 //! least one block of elements created. Note: 0 is a possible value
 //! in this case no blocks are allocated. Default value is 1.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-IContainer<T,N,Block>::IContainer(const uint32_t nb_elts)
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+IContainer<T,N,Block>::IContainer(const size_t nb_elts)
 {
   m_stored_elements = 0;
   m_allocated_blocks = 0;
@@ -52,11 +52,11 @@ IContainer<T,N,Block>::IContainer(const uint32_t nb_elts)
 //! pre-allocate. This value is automaticly rounded up to have at
 //! least one block of elements created.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-void IContainer<T,N,Block>::reserve(const uint32_t nb_elts)
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+void IContainer<T,N,Block>::reserve(const size_t nb_elts)
 {
-  std::cout << "Reserving (" << nb_elts << " + " << M << " - 1) >> " << N << " blocks\n";
+  std::cout << "Reserving (" << nb_elts << " + " << (int) M << " - 1) >> " << (int) N << " blocks\n";
   reserveBlocks((nb_elts + M - 1) >> N);
 }
 
@@ -64,15 +64,15 @@ void IContainer<T,N,Block>::reserve(const uint32_t nb_elts)
 //! \param nb_blocks is the number of blocks of elements of type T
 //! we want to pre-allocate.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-void IContainer<T,N,Block>::reserveBlocks(const uint32_t nb_blocks)
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+void IContainer<T,N,Block>::reserveBlocks(const size_t nb_blocks)
 {
   std::cout << "Reserving " << nb_blocks << " blocks\n";
-  assert(nb_blocks < 16U);
+  assert(nb_blocks < 16ul);
   m_blocks.reserve(nb_blocks);
 
-  uint32_t i = nb_blocks;
+  size_t i = nb_blocks;
   while (i--)
     {
       m_blocks.push_back(newBlock());
@@ -89,17 +89,17 @@ void IContainer<T,N,Block>::reserveBlocks(const uint32_t nb_blocks)
 //! the position of the last insereted element) or if the element
 //! has been removed.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-T& IContainer<T,N,Block>::get(const uint32_t nth)
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+T& IContainer<T,N,Block>::get(const size_t nth)
 {
   if (outofbound(nth))
     {
       throw std::out_of_range("Out of range index " + std::to_string(nth));
     }
 
-  const uint32_t id = nth >> N;
-  const uint32_t sid = MODULO(nth, M);
+  const size_t id = nth >> N;
+  const size_t sid = MODULO(nth, M);
 
   if (!IS_OCCUPIED(id, sid))
     {
@@ -111,17 +111,17 @@ T& IContainer<T,N,Block>::get(const uint32_t nth)
 // **************************************************************
 // FIXME code in double with previous code
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-T const& IContainer<T,N,Block>::get(const uint32_t nth) const
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+T const& IContainer<T,N,Block>::get(const size_t nth) const
 {
   if (outofbound(nth))
     {
       throw std::out_of_range("Out of range index " + std::to_string(nth));
     }
 
-  const uint32_t id = nth >> N;
-  const uint32_t sid = MODULO(nth, M);
+  const size_t id = nth >> N;
+  const size_t sid = MODULO(nth, M);
 
   if (!IS_OCCUPIED(id, sid))
     {
@@ -133,17 +133,17 @@ T const& IContainer<T,N,Block>::get(const uint32_t nth) const
 // **************************************************************
 //
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-void IContainer<T,N,Block>::modify(const uint32_t nth, T const& e) const
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+void IContainer<T,N,Block>::modify(const size_t nth, T const& e) const
 {
   if (outofbound(nth))
     {
       throw std::out_of_range("Out of range index " + std::to_string(nth));
     }
 
-  const uint32_t id = nth >> N;
-  const uint32_t sid = MODULO(nth, M);
+  const size_t id = nth >> N;
+  const size_t sid = MODULO(nth, M);
 
   if (!IS_OCCUPIED(id, sid))
     {
@@ -156,12 +156,12 @@ void IContainer<T,N,Block>::modify(const uint32_t nth, T const& e) const
 // **************************************************************
 //
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
 const T& IContainer<T,N,Block>::operator[](size_t nth) const
 {
-  const uint32_t id = nth >> N;
-  const uint32_t sid = MODULO(nth, M);
+  const size_t id = nth >> N;
+  const size_t sid = MODULO(nth, M);
 
   return m_blocks[id]->m_block[sid];
 }
@@ -169,12 +169,12 @@ const T& IContainer<T,N,Block>::operator[](size_t nth) const
 // **************************************************************
 //
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
 T& IContainer<T,N,Block>::operator[](size_t nth)
 {
-  const uint32_t id = nth >> N;
-  const uint32_t sid = MODULO(nth, M);
+  const size_t id = nth >> N;
+  const size_t sid = MODULO(nth, M);
 
   return m_blocks[id]->m_block[sid];
 }
@@ -184,17 +184,17 @@ T& IContainer<T,N,Block>::operator[](size_t nth)
 //! to check.
 //! \return true if an element is present, else return false.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-bool IContainer<T,N,Block>::occupied(const uint32_t nth) const
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+bool IContainer<T,N,Block>::occupied(const size_t nth) const
 {
   if (outofbound(nth))
     {
       throw std::out_of_range("Out of range index " + std::to_string(nth));
     }
 
-  const uint32_t id = nth >> N;
-  const uint32_t sid = MODULO(nth, M);
+  const size_t id = nth >> N;
+  const size_t sid = MODULO(nth, M);
 
   return !(!(IS_OCCUPIED(id, sid)));
 }
@@ -203,17 +203,17 @@ bool IContainer<T,N,Block>::occupied(const uint32_t nth) const
 //! \param nth the n'th element (index) of the container we want
 //! to be not empty.
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
-void IContainer<T,N,Block>::occupy(const uint32_t nth)
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
+void IContainer<T,N,Block>::occupy(const size_t nth)
 {
   if (outofbound(nth))
     {
       throw std::out_of_range("Out of range index " + std::to_string(nth));
     }
 
-  const uint32_t id = nth >> N;
-  const uint32_t sid = MODULO(nth, M);
+  const size_t id = nth >> N;
+  const size_t sid = MODULO(nth, M);
 
   if (!IS_OCCUPIED(id, sid))
     {
@@ -225,12 +225,12 @@ void IContainer<T,N,Block>::occupy(const uint32_t nth)
 // **************************************************************
 //!
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
 void IContainer<T,N,Block>::garbage()
 {
   // Important: start by the end
-  uint32_t index = m_allocated_blocks;
+  size_t index = m_allocated_blocks;
   while (index--)
     {
       ContainerBitField subindex = M;
@@ -252,8 +252,8 @@ void IContainer<T,N,Block>::garbage()
 // **************************************************************
 //!
 // **************************************************************
-template<typename T, const uint32_t N,
-         template<typename X, const uint32_t Y> class Block>
+template<typename T, const size_t N,
+         template<typename X, const size_t Y> class Block>
 void IContainer<T,N,Block>::debug() const
 {
   if (empty())
@@ -261,12 +261,12 @@ void IContainer<T,N,Block>::debug() const
       std::cout << "The container is empty" << std::endl;
       return ;
     }
-  for (uint32_t index = 0; index < m_allocated_blocks; ++index)
+  for (size_t index = 0ul; index < m_allocated_blocks; ++index)
     {
       std::cout << "Block " << index << ": " << std::endl;
 
       // Show elements stored
-      for (uint32_t subindex = 0; subindex < M; ++subindex)
+      for (size_t subindex = 0ul; subindex < M; ++subindex)
         {
           if (IS_OCCUPIED(index, subindex))
             {
